requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
tqdm==4.66.4
seaborn==0.13.2
matplotlib==3.8.0
nltk==3.9.0
textblob==0.15.3
spacy==3.6.0
numpy==1.26.0 
